+++ tests/tests_integration/test_api/test_data_modeling/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+
+from cognite.client import CogniteClient
+from cognite.client.data_classes.data_modeling import Space
+
+
+@pytest.fixture(scope="session")
+def integration_test_space(cognite_client: CogniteClient) -> Space:
+    return cognite_client.data_modeling.spaces.retrieve("IntegrationTestsImmutable")