+++ tests/tests_integration/test_api/test_data_modeling/test_spaces.py
@@ -2,19 +2,19 @@ from __future__ import annotations
 
 import pytest
 
+import cognite.client.data_classes.data_modeling as models
 from cognite.client import CogniteClient
-from cognite.client.data_classes import Space, SpaceList
 
 
 @pytest.fixture(scope="function")
 def cdf_spaces(cognite_client):
     spaces = cognite_client.data_modeling.spaces.list(limit=-1)
     assert len(spaces) > 0, "Please create at least one space in CDF."
-    yield spaces
+    return spaces
 
 
-def _dump(list_: SpaceList | Space) -> list[dict]:
-    if isinstance(list_, Space):
+def _dump(list_: models.SpaceList | models.Space) -> list[dict]:
+    if isinstance(list_, models.Space):
         output = [list_.dump()]
     else:
         output = sorted((s.dump() for s in list_), key=lambda s: s["space"])
@@ -25,14 +25,14 @@ def _dump(list_: SpaceList | Space) -> list[dict]:
 
 
 class TestSpacesAPI:
-    def test_list(self, cognite_client: CogniteClient, cdf_spaces: SpaceList):
+    def test_list(self, cognite_client: CogniteClient, cdf_spaces: models.SpaceList):
         actual_space_in_cdf = cognite_client.data_modeling.spaces.list(limit=-1)
 
         assert _dump(actual_space_in_cdf) == _dump(cdf_spaces)
 
     def test_create_retrieve_and_delete(self, cognite_client: CogniteClient):
         # Arrange
-        my_space = Space(
+        my_space = models.SpaceApply(
             space="myNewSpace", name="My New Space", description="This is part of the integration testing for the SDK."
         )
 
@@ -42,9 +42,7 @@ class TestSpacesAPI:
 
         # Assert
         assert retrieved_space.dump() == created_space.dump()
-        expected = retrieved_space.dump()
-        expected.pop("created_time")
-        expected.pop("last_updated_time")
+        expected = retrieved_space.as_apply().dump()
         assert my_space.dump() == expected
 
         # Act
@@ -54,8 +52,8 @@ class TestSpacesAPI:
         assert deleted_space == my_space.space
         assert cognite_client.data_modeling.spaces.retrieve(space=my_space.space) is None
 
-    def test_retrieve_multiple(self, cognite_client: CogniteClient, cdf_spaces: SpaceList):
-        retrieved_spaces = cognite_client.data_modeling.spaces.retrieve_multiple([s.space for s in cdf_spaces])
+    def test_retrieve_multiple(self, cognite_client: CogniteClient, cdf_spaces: models.SpaceList):
+        retrieved_spaces = cognite_client.data_modeling.spaces.retrieve([s.space for s in cdf_spaces])
 
         assert _dump(retrieved_spaces) == _dump(cdf_spaces)